finished and test listing controller

+++ src/main/java/ZERL/server/services/ListingService.java
@@ -30,7 +30,7 @@ public class ListingService {
         return listingRepository.updateListing(listing);
     }
 
-    public Listing deleteListing(double id) {
+    public Listing deleteListing(int id) {
         return listingRepository.deleteListing(id);
     }
 }
